# kaseki-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kaseki Agent API",
    "version": "1.49.0",
    "description": "Ephemeral coding-agent runner: orchestrates Pi CLI via Docker for automated code modifications with validation and deployment",
    "contact": {
      "name": "CyanAutomation",
      "url": "https://github.com/CyanAutomation/kaseki-agent"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/CyanAutomation/kaseki-agent/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://kaseki.example.com",
      "description": "Production server (configure as needed)"
    }
  ],
  "paths": {
...
    "/api/runs": {
      "post": {
        "operationId": "triggerRun",
        "summary": "Trigger a new kaseki run",
        "description": "Submits a new job to the queue. Returns 202 Accepted with job metadata. Use the ID to poll status.",
        "tags": [
          "Run Management"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "repoUrl"
                ],
                "properties": {
                  "repoUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "Git repository URL"
                  },
                  "ref": {
                    "type": "string",
                    "default": "main",
                    "description": "Git branch/tag/commit"
                  },
                  "taskPrompt": {
                    "type": "string",
                    "description": "Task prompt for Pi agent"
                  },
                  "changedFilesAllowlist": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "File patterns to allow changes in"
                  },
                  "maxDiffBytes": {
                    "type": "integer",
                    "description": "Max diff size in bytes"
                  },
                  "validationCommands": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Validation commands to run"
                  },
                  "scouting": {
                    "type": "object",
                    "description": "Pre-coding Pi scouting controls",
                    "properties": {
                      "enabled": {
                        "type": "boolean",
                        "description": "Enable the pre-coding Pi scouting phase"
                      },
                      "model": {
                        "type": "string",
                        "description": "Optional Pi model override for scouting"
                      },
                      "timeoutSeconds": {
                        "type": "integer",
                        "minimum": 60,
                        "maximum": 10800,
                        "description": "Optional scouting timeout in seconds"
                      }
                    }
                  },
                  "goalCheck": {
                    "type": "object",
                    "description": "Post-coding goal-check evaluator controls",
                    "properties": {
                      "enabled": {
                        "type": "boolean",
                        "description": "Enable the post-validation goal-check Pi evaluator"
                      },
                      "maxRetries": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 5,
                        "description": "Maximum coding-agent retries after goal-check misses"
                      },
                      "model": {
                        "type": "string",
                        "description": "Optional Pi model override for goal checking"
                      },
                      "timeoutSeconds": {
                        "type": "integer",
                        "minimum": 60,
                        "maximum": 10800,
                        "description": "Optional goal-check timeout in seconds"
                      }
                    }
                  },
                  "taskMode": {
                    "type": "string",
                    "enum": [
                      "patch",
                      "inspect"
                    ],
                    "description": "Task mode"
                  },
                  "publishMode": {
                    "type": "string",
                    "enum": [
                      "auto",
                      "none",
                      "branch",
                      "pr",
                      "draft_pr"
                    ],
                    "description": "Publishing mode after validation: pr creates a normal pull request (controller default when omitted), draft_pr creates a draft pull request, branch pushes only, auto publishes when credentials are available and skips if missing, none skips publishing"
                  },
                  "skipPreAgentValidation": {
                    "type": "boolean",
                    "description": "Skip baseline validation before the Pi agent starts. Useful for fast inspect-only runs; validation still runs after agent execution unless otherwise configured."
                  },
                  "startupCheck": {
                    "type": "boolean",
                    "description": "Start a worker container and exit after boot checks"
                  },
                  "startupCheckMode": {
                    "type": "string",
                    "enum": [
                      "boot",
                      "baseline-validation"
                    ],
                    "description": "Startup check depth: boot-only container smoke test or baseline validation dry-run"
                  },
                  "webhookConfig": {
                    "type": "object",
                    "description": "Webhook configuration for job events"
                  },
                  "tracing": {
                    "type": "object",
                    "description": "Request tracing identifiers"
                  },
                  "idempotencyKey": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Idempotency key for safe retries"
                  },
                  "timeoutSeconds": {
                    "type": "integer",
                    "minimum": 60,
                    "maximum": 10800,
                    "description": "Per-run timeout in seconds"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotency replay: job already exists with same configuration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "status",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique kaseki instance ID (kaseki-N)",
                      "example": "kaseki-42"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "completed",
                        "failed"
                      ],
                      "description": "Current job status"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp when the job was created"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for tracing"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Unique request ID"
                    },
                    "cached": {
                      "type": "boolean",
                      "description": "True when returned from idempotency replay"
                    },
                    "completedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp when the job completed"
                    },
                    "exitCode": {
                      "type": "integer",
                      "description": "Exit code from kaseki execution"
                    },
                    "failureClass": {
                      "type": "string",
                      "description": "Classification of failure (e.g., validation_failed, timeout)"
                    },
                    "error": {
                      "type": "string",
                      "description": "Error message if the job failed"
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Job accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "status",
                    "createdAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique kaseki instance ID (kaseki-N)",
                      "example": "kaseki-42"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "running",
                        "completed",
                        "failed"
                      ],
                      "description": "Current job status"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp when the job was created"
                    },
                    "correlationId": {
                      "type": "string",
                      "description": "Request correlation ID for tracing"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Unique request ID"
                    },
                    "cached": {
                      "type": "boolean",
                      "description": "True when returned from idempotency replay"
                    },
                    "completedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "ISO 8601 timestamp when the job completed"
                    },
                    "exitCode": {
                      "type": "integer",
                      "description": "Exit code from kaseki execution"
                    },
                    "failureClass": {
                      "type": "string",
                      "description": "Classification of failure (e.g., validation_failed, timeout)"
                    },
                    "error": {
                      "type": "string",
                      "description": "Error message if the job failed"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listRuns",
        "summary": "List all runs",
        "description": "Returns paginated list of all kaseki runs, newest first. Includes basic metadata for each run.",
        "tags": [
          "Run Management"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 500
            },
            "description": "Maximum number of runs to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "runs",
                    "total"
                  ],
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "status",
                          "createdAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Unique kaseki instance ID (kaseki-N)",
                            "example": "kaseki-42"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "queued",
                              "running",
                              "completed",
                              "failed"
                            ],
                            "description": "Current job status"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "ISO 8601 timestamp when the job was created"
                          },
                          "correlationId": {
                            "type": "string",
                            "description": "Request correlation ID for tracing"
                          },
                          "requestId": {
                            "type": "string",
                            "description": "Unique request ID"
                          },
                          "cached": {
                            "type": "boolean",
                            "description": "True when returned from idempotency replay"
                          },
                          "completedAt": {
                            "type": "string",
                            "format": "date-time",
                            "description": "ISO 8601 timestamp when the job completed"
                          },
                          "exitCode": {
                            "type": "integer",
                            "description": "Exit code from kaseki execution"
                          },
                          "failureClass": {
                            "type": "string",
                            "description": "Classification of failure (e.g., validation_failed, timeout)"
                          },
                          "error": {
                            "type": "string",
                            "description": "Error message if the job failed"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total number of runs available"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/status": {
      "get": {
        "operationId": "getRunStatus",
        "summary": "Poll run status",
        "description": "Returns current status of a kaseki run, including progress, elapsed time, and timeout risk percentage.",
        "tags": [
          "Run Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^kaseki-\\d+$"
            },
            "description": "Kaseki instance ID (e.g., kaseki-42)"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current run status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/cancel": {
      "post": {
        "operationId": "cancelRun",
        "summary": "Cancel a run",
        "description": "Cancels a queued or running kaseki job. Returns 200 if cancellation was accepted; operation may complete asynchronously.",
        "tags": [
          "Run Management"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^kaseki-\\d+$"
            },
            "description": "Kaseki instance ID"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Run not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    },
                    "requestId": {
                      "type": "string",
                      "description": "Request ID for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "id",
          "status",
          "elapsedSeconds",
          "timeoutRiskPercent"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Kaseki instance ID"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ]
          },
          "progress": {
            "type": "object",
            "properties": {
              "stage": {
                "type": "string",
                "description": "Current stage name"
              },
              "percentComplete": {
                "type": "integer",
                "minimum": 0,
                "maximum": 100,
                "description": "Progress percentage (0-100)"
              },
              "message": {
                "type": "string",
                "description": "Detailed progress message"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "description": "ISO 8601 timestamp of last update"
              }
            }
          },
          "elapsedSeconds": {
            "type": "number",
            "description": "Elapsed time since job start in seconds"
          },
          "timeoutRiskPercent": {
            "type": "number",
            "description": "Percentage of timeout used (0-100+); values >85 indicate imminent timeout"
          },
          "exitCode": {
            "type": "integer",
            "description": "Exit code (only if completed)"
          },
          "failureClass": {
            "type": "string",
            "description": "Failure classification (only if failed)"
          },
          "error": {
            "type": "string",
            "description": "Error message (only if failed)"
          },
          "resultDir": {
            "type": "string",
            "description": "Path to results directory on server"
          },
          "resultSummaryContent": {
            "type": "string",
            "description": "Human-readable markdown summary (truncated to 64KB)"
          },
          "failureJsonContent": {
            "type": "object",
            "description": "Structured failure information (only if failed)"
          },
          "artifacts": {
            "type": "object",
            "properties": {
              "metadataJson": {
                "type": "boolean"
              },
              "analysisMd": {
                "type": "boolean"
              },
              "resultSummaryMd": {
                "type": "boolean"
              },
              "failureJson": {
                "type": "boolean"
              },
              "stderrLog": {
                "type": "boolean"
              },
              "availableFiles": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "List of available artifact files"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "Bearer token authentication. Provide your API key (from KASEKI_API_KEYS environment variable) as the bearer token."
      }
    }
  },
...
}